Remove Crystime::Span
This is a speculation that Crystime::Span does not need to keep track
of @ts itself, and that it therefore does not need to exist - Time::Span
can be used directly.

The thing to resolve here is what happens when a person does:

vd2 - vd1

(In Crystal, Time - Time results in Time::Span, while Time + Time is
not supported.)

Since VirtualDates can contain complex rules (and some are addable/
subtractable), removing Crystime::Span will remove the ability to
add/subtract two complex VDs.

It remains to be seen if this is OK, or if Crystime::Span needs to be
brought back.